Fix implicit narrowing in tf2 time conversion (#1834)
Use explicit fixed-width casts when constructing rclcpp::Time, matching createHeader and silencing MSVC C4244 warnings.

Generated-by: OpenAI Codex (GPT-5)

+++ rviz_common/src/rviz_common/transformation/tf2_helpers/tf2_conversion_helpers.cpp
@@ -59,7 +59,9 @@ fromTf2TimePoint(const tf2::TimePoint & tf2_time)
   const auto seconds = std::chrono::time_point_cast<std::chrono::seconds>(tf2_time);
   const auto nanoseconds = std::chrono::time_point_cast<std::chrono::nanoseconds>(tf2_time) -
     std::chrono::time_point_cast<std::chrono::nanoseconds>(seconds);
-  return rclcpp::Time(seconds.time_since_epoch().count(), nanoseconds.count());
+  return rclcpp::Time(
+    static_cast<std::int32_t>(seconds.time_since_epoch().count()),
+    static_cast<std::uint32_t>(nanoseconds.count()));
 }
 
 tf2::TimePoint
